Add a plugin to enable irony-mode on a CMake project.
This is done by generating and parsing compile_commands.json (see
cmake -DCMAKE_EXPORT_COMPILE_COMMANDS=ON <args...>).

The implementation is a bit ugly, a lot of assumptions and some
potential errors and limitations.

BUT it seems to work sufficiently well and is "so easy" to setup, I'm
quite happy with it.