remove check compile and server lookup when irony-mode is enabled
See https://github.com/Sarcasm/irony-mode/issues/105

The irony-mode setup code was doing too many costly operations even if a
buffer was just opened and closed very fast. In this situation irony was
looking for the server executable and triggered a "check-compile". When
semantic is enabled it parses the system headers used by the project,
which mean many C++ files were opened, parsed and closed very quickly.
Irony-mode was checking all these file and this made Emacs barely
usable, irony-server was eating a lot of memory. Now irony-server is
looked up "lazily" and the buffer is no longer checked when opened, this
latter feature may come back in a future commit but will happen only in
buffers where the users is actively doing something.