Fix the CMake files to work on Windows
LibClangDiagnosticsChecker.c, was using C instead of C++. On Windows it
means we were compiling it in C89 mode or so, a mode where declaring
variables was not possible in for-loops, ...Just use C++ instead, it
should be less fragile.

This changes only take into consideration a native build, it hasn't been
tested on an emulation layer such as mingw.

We make the assumption that:
- a Windows compiler is used
- the package that provides libclang is the one from http://llvm.org/releases/
- libclang.dll is available in the PATH (or at least in emacs's
  exec-path).

Also remove unnecessary HINTS variables LIBCLANG_INCLUDE_PATH and
LIBCLANG_LIBRARY_PATH, the user can just use LIBCLANG_INCLUDE_DIR and
LIBCLANG_LIBRARY.