irony-server: simplify getCompileCommands implementation
I'm not against new files but right now it was more straightforward to
follow the pattern of the other methods of the class Irony.

Everything is implemented in one function when possible. I don't see
these functions growing very much in the future so we don't need extra
abstraction over the libclang API.

Implementing it that way make it better performance-wise. At least in
terms of memory allocation, we reduce them to the minimum by not
collecting the data into some vector/string and then display them.

Fixed a bug in the call to clang_CompileCommands_getCommand(), 0 was
always used instead of 'i'.