irony-server: rework support::quoted() support for temporaries
I have a doubt, whether or not temporary objects are well supported by
the implementation or not. Since I'm not sure and have difficulty to
find a correct answer I prefer to go the safe way and make a copy of the
passed string.

I will be happy to revise this but this is unrelated to the compilation
database so that can be discussed in its own PR/issue.