irony: use start-process-shell-command instead of start-process
Clang uses fprintf(stderr, ) to print some stuff and maybe
std::cerr (?). It is annoying and error prone for me to redirect both C
and C++ streams. Redirecting clog has already proven to be a source of
error. Since the shell handle this nicely with '2> stderr.log', both on
Windows and "Unix", I think it's better to do keep it simple for now.

I tested on big input on both Linux and Windows, there is no impact with
the intermediate shell contrary to the hypothesis evoked here:
https://github.com/Sarcasm/irony-mode/issues/131#issuecomment-58940244.