ignore errors in server response processing
In function `irony--server-process-filter`, when multiple responses are
returned, will call `irony--process-server-response` to process theme on
by one. But if some one of the responses (not the last) signals error in
processing, then `irony--server-process-filter` will be interrupted,
that will leave some responses not processed, and corresponding
callbacks in the process not poped. Then follow-up requests will call
wrong callbacks and waiting in `accept-process-output` forever.

Run `irony-server-kill` manually will clear the callbacks and fix this,
but ignore errors in server response processing will be better.