irony-server: Adds keep going flag to translation unit parse options.
For versions libclang versions that have it, this adds
CXTranslationUnit_KeepGoing to the translation unit parse options.

Without it, it will stop parsing on errors, which means that for instance
unused variables will not be reported until the error has been fixed.

This also changes the existing CINDEX_VERSION check in TU manager to
a preprocessor check, as it's used in other files.