irony-server: Rewrites dumpDiagnostics(); adds end column, notes and hints.
This rewrites dumpDiagnostics() so that it also finds the end column,
includes notes and fixit hints.

The output format has not changed, but instead of outputting the offset,
it now outputs the end column. This field was previously unused, but
a new function as been added to irony-diagnostics.el to pick it up.

This rewrite has two goals; find the end column to speed up Flycheck's
error-region for flycheck-irony (separate project and change) and improve
diagnostic reporting by including notes and fixit hints.

Notes are kept in child diagnostics, so these are now picked up and
parsed as well.

Once parsed, notes are collapsed into their parent if on the same line
(since it's mostly extra information about the parent).

Some more filtering is also done to get rid of reference to other files
(which are irrelevant for flycheck-irony) and "delcared here" notes
that will appear out of context before the actual message (usually) when
marked by Flycheck.

Messages are changed to also include notes (or'ed together, which seems
to work quite well) and a hint, if there is only one and it's not attached
to a note (which already spells out what the hint says).

Explanatory comments have been added in the code to clarify what it's
trying to achieve.