irony-cdb-json: fix compiler flags when using ccache
This commit changes the cleanup of the compiler flags to call `irony-cdb--remove-compiler-from-flags`. This function is used by the `libclang` database to properly remove the compiler from the flags. In particular: it takes into account if the compiler is wrapped by a call to `ccache`.

Similar changes might be needed in other databases to be able to properly support ccache.