CMake: remove unnecessary cruft
These things are sometimes trying to be helpful,
but they clutter the CMakeLists.txt,
while things like -DCMAKE_CXX_FLAGS='...' can be specified by the user when he
needs customization.

Same for in-source builds, we are trying to be helpful,
but it's not actually an issue.
If in-source builds are to forbid, it should be CMake to decide, not us.